Bugfix TPMMS in SPIDER
Previously a newline character was inserted no matter
if a value has been written or skipped.
A newline character should only be inserted if the
value was not skipped. If a newline appears at the
end of the file makes no difference to the reading
procedures used in the TPMMS component.